Trigger Binder to build image on pushes to master
Use selenium to trigger Binder to have repo2docker start the build of
the image anytime a new push to master occurs. This ensures that there
will always be a recently built Binder example so that if someone clicks
on the launch Binder badge they don't have to wait for it to build.
selenium is used instead of just `python -m webbrowser` as the browser
needs to be open on the URL long enough for Binder to trigger.