Read Double32_t values with truncated mantissa
This commit implements unpacking of the Double32_t format in which the exponent
and mantissa of the original double value (cast to single precision) are
streamed as 8-bit and 16-bit integers respectively.

Depending on the specification provided by the user the ROOT's Double32_t
values can be packed and streamed to a ROOT file in other two formats. In the
first mode the original double precision values are simply cast to a single
precision and streamed without further manipulation. In the second mode the
double values (again, cast to float and scaled to a given range) are saved as
32-bit integers. These two modes are already taken care of by the
`uproot.interp.auto.interpret()` function and the
`uproot.interp.numerical.asdouble32` interpretation.

For details on Double32_t streaming modes see
https://root.cern/doc/v616/classTBufferFile.html#a44c2adb6fb1194ec999b84aed259e5bc