Add connection period and duration limits to Binder build API curl
The use of connect-timeout is to make sure that the curl doesn't try to
repeatedly hit the URL if the connection is down or unresponsive for an
unreasonably long time. The use of max-time is so that for longer builds
the curl doesn't try to follow them for the duration. It just hits the
API endpoint and then exits at most 60 seconds later. Additionally the
trigger function accepts --max-time exit return code, 28, as valid and
then return success.

c.f.
https://unix.stackexchange.com/questions/94604/does-curl-have-a-timeout/94612