Get fixed size array dimension from leaf description
Extracting data dimension from the leaf's title does not always work for
sub-branches, i.e. when the array is an element of a parrent jagged array. In
such case we try to use the dimension saved in the leaf's description. This
fix was not checked against multidimensional arrays but it does work for the
case introduced in the Event.root test.
See 2041da93ecb6b7df4b0c0d1924dfb95a3460a541 and 862dc72789cb74289f6fffebcf2f4b87a87d75fa