Trigger Binder builds on both GKE and OVH clusters
Given that the Binder Federation exists on both GKE and OVH there is no
guarantee that the cluster that a build is triggered on is the same
cluster that a user will be directed to when they launch the Binder. As
a result, both clusters should be triggered when a PR is merged.

c.f. https://discourse.jupyter.org/t/the-binder-federation/1286